requirements.txt
flask==3.0.2
flask-sqlalchemy==3.1.1
python-dotenv==1.0.1
flask-migrate==4.0.5
APScheduler==3.10.4
Werkzeug==3.0.6
Jinja2==3.1.5
click==8.1.8
itsdangerous==2.2.0
SQLAlchemy==2.0.38
alembic==1.14.1
blinker==1.8.2
importlib-metadata==8.5.0
MarkupSafe==2.1.5
typing-extensions==4.12.2
colorama==0.4.6
pytz==2025.1
six==1.17.0
tzlocal==5.2
tzdata==2025.1
pyppeteer==1.0.2
python-dateutil==2.8.2
WTForms==3.1.2